auth/requirements.txt
torch
opencv-python
face_recognition
numpy
Pillow
scikit-learn
pyserial
matplotlib